test(slatetohtml): reorganise Reorganise for https://github.com/thompsonsj/slate-serializers-demo/pull/32
